Refactor OptiGlue 1.18.2
