+++ OptiGlue/1.18.2/src/main/kotlin/opekope2/optiglue_1_18_2/RegistryLookupServiceImpl.kt
@@ -1,4 +1,4 @@
-package opekope2.optiglue.mc_1_18_2
+package opekope2.optiglue_1_18_2
 
 import net.minecraft.block.Block
 import net.minecraft.entity.Entity
